+++ src/app/app.module.ts
@@ -1,6 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { Router } from '@angular/router';
+import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -31,6 +32,7 @@ import { LogoutComponent } from './layout/components';
     ProductsModule,
     CartModule,
     SharedModule,
+    HttpClientModule,
     OrdersModule,
     AppRoutingModule
   ],